// @flywave/flywave-topo/src/clipping/clip-plane.d.ts
export interface ClipPlaneProps {
    normal?: XYZProps;
    dist?: number;
    invisible?: boolean;
    interior?: boolean;
}
export declare class ClipPlane extends Plane3d implements Clipper, PolygonClipper {
    private _inwardNormal;
    private _distanceFromOrigin;
    private _invisible;
    private _interior;
    private constructor();
    isAlmostEqual(other: ClipPlane): boolean;
    clone(): ClipPlane;
    cloneNegated(): ClipPlane;
    static createPlane(plane: Plane3dByOriginAndUnitNormal, invisible?: boolean, interior?: boolean, result?: ClipPlane): ClipPlane;
    static createNormalAndDistance(normal: Vector3d, distance: number, invisible?: boolean, interior?: boolean, result?: ClipPlane): ClipPlane | undefined;
    static createNormalAndPoint(normal: Vector3d, point: Point3d, invisible?: boolean, interior?: boolean, result?: ClipPlane): ClipPlane | undefined;
    static createOriginAndVectors(origin: Point3d, vectorA: Vector3d, vectorB: Vector3d, invisible?: boolean, interior?: boolean, result?: ClipPlane): ClipPlane | undefined;
    static createNormalAndPointXYZXYZ(normalX: number, normalY: number, normalZ: number, originX: number, originY: number, originZ: number, invisible?: boolean, interior?: boolean, result?: ClipPlane): ClipPlane | undefined;
    toJSON(): ClipPlaneProps;
    static fromJSON(json: ClipPlaneProps, result?: ClipPlane): ClipPlane | undefined;
    setFlags(invisible: boolean, interior: boolean): void;
    get distance(): number;
    get inwardNormalRef(): Vector3d;
    get interior(): boolean;
    get invisible(): boolean;
    static createEdgeAndUpVector(point0: Point3d, point1: Point3d, upVector: Vector3d, tiltAngle?: Angle, result?: ClipPlane): ClipPlane | undefined;
    static createEdgeXY(point0: Point3d, point1: Point3d, result?: ClipPlane): ClipPlane | undefined;
    getPlane3d(): Plane3dByOriginAndUnitNormal;
    getPlane4d(): Point4d;
    setPlane4d(plane: Point4d): void;
    weightedAltitude(point: Point4d): number;
    altitude(point: Point3d): number;
    altitudeXYZ(x: number, y: number, z: number): number;
    normalX(): number;
    normalY(): number;
    normalZ(): number;
    velocity(vector: Vector3d): number;
    velocityXYZ(x: number, y: number, z: number): number;
    dotProductPlaneNormalPoint(point: Point3d): number;
    isPointOnOrInside(spacePoint: Point3d, tolerance?: number): boolean;
    isPointInside(point: Point3d, tolerance?: number): boolean;
    isPointOn(point: Point3d, tolerance?: number): boolean;
    appendIntersectionRadians(arc: Arc3d, intersectionRadians: GrowableFloat64Array): void;
    private static readonly _clipArcFractionArray;
    announceClippedArcIntervals(arc: Arc3d, announce?: AnnounceNumberNumberCurvePrimitive): boolean;
    getBoundedSegmentSimpleIntersection(pointA: Point3d, pointB: Point3d): number | undefined;
    transformInPlace(transform: Transform): boolean;
    setInvisible(invisible: boolean): void;
    negateInPlace(): void;
    offsetDistance(offset: number): void;
    clipConvexPolygonInPlace(xyz: GrowableXYZArray, work: GrowableXYZArray, inside?: boolean, tolerance?: number): number;
    multiplyPlaneByMatrix4d(matrix: Matrix4d, invert?: boolean, transpose?: boolean): boolean;
    announceClippedSegmentIntervals(f0: number, f1: number, pointA: Point3d, pointB: Point3d, announce?: (fraction0: number, fraction1: number) => void): boolean;
    getFrame(): Transform;
    intersectRange(range: Range3d, addClosurePoint?: boolean): GrowableXYZArray | undefined;
    appendPolygonClip(xyz: IndexedXYZCollection, insideFragments: GrowableXYZArray[], outsideFragments: GrowableXYZArray[], arrayCache: GrowableXYZArrayCache): void;
    projectPointToPlane(spacePoint: Point3d, result?: Point3d): Point3d;
}