// @flywave/flywave-topo/src/clipping/convex-clip-plane-set.d.ts
export type ConvexClipPlaneSetProps = ClipPlaneProps[];
export declare class ConvexClipPlaneSet implements Clipper, PolygonClipper {
    static readonly hugeVal = 1e+37;
    private readonly _planes;
    private constructor();
    toJSON(): ConvexClipPlaneSetProps;
    static fromJSON(json: ConvexClipPlaneSetProps | undefined, result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    isAlmostEqual(other: ConvexClipPlaneSet): boolean;
    static createPlanes(planes: Array<ClipPlane | Plane3dByOriginAndUnitNormal>, result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    static createRange3dPlanes(range: Range3d, lowX?: boolean, highX?: boolean, lowY?: boolean, highY?: boolean, lowZ?: boolean, highZ?: boolean): ConvexClipPlaneSet;
    static createEmpty(result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    negateAllPlanes(): void;
    static createXYBox(x0: number, y0: number, x1: number, y1: number, result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    static createXYPolyLine(points: Point3d[], interior: boolean[] | undefined, leftIsInside: boolean, result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    static createXYPolyLineInsideLeft(points: Point3d[], result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    static setPlaneAndXYLoopCCW(points: GrowableXYZArray, planeOfPolygon: ClipPlane, frustum: ConvexClipPlaneSet): void;
    clone(result?: ConvexClipPlaneSet): ConvexClipPlaneSet;
    get planes(): ClipPlane[];
    hasIntersectionWithRay(ray: Ray3d, result?: Range1d, tolerance?: number): boolean;
    multiplyPlanesByMatrix4d(matrix: Matrix4d, invert?: boolean, transpose?: boolean): boolean;
    isPointInside(point: Point3d): boolean;
    isPointOnOrInside(point: Point3d, tolerance?: number): boolean;
    isSphereInside(centerPoint: Point3d, radius: number): boolean;
    announceClippedSegmentIntervals(f0: number, f1: number, pointA: Point3d, pointB: Point3d, announce?: AnnounceNumberNumber): boolean;
    private static readonly _clipArcFractionArray;
    announceClippedArcIntervals(arc: Arc3d, announce?: AnnounceNumberNumberCurvePrimitive): boolean;
    clipUnboundedSegment(pointA: Point3d, pointB: Point3d, announce?: AnnounceNumberNumber): boolean;
    transformInPlace(transform: Transform): void;
    clipConvexPolygonInPlace(xyz: GrowableXYZArray, work: GrowableXYZArray, tolerance?: number): void;
    clipInsidePushOutside(xyz: IndexedXYZCollection, outsideFragments: GrowableXYZArray[] | undefined, arrayCache: GrowableXYZArrayCache): GrowableXYZArray | undefined;
    classifyPointContainment(points: Point3d[], onIsOutside: boolean): ClipPlaneContainment;
    static createSweptPolyline(points: Point3d[], upVector: Vector3d, tiltAngle?: Angle): ConvexClipPlaneSet | undefined;
    addPlaneToConvexSet(plane: ClipPlane | Plane3dByOriginAndUnitNormal | undefined): void;
    clipPointsOnOrInside(points: Point3d[], inOrOn: Point3d[], out: Point3d[]): void;
    polygonClip(input: GrowableXYZArray | Point3d[], output: GrowableXYZArray, work: GrowableXYZArray, planeToSkip?: ClipPlane): void;
    reloadSweptPolygon(points: Point3d[], sweepDirection: Vector3d, sideSelect: number): number;
    computePlanePlanePlaneIntersections(points: Point3d[] | undefined, rangeToExtend: Range3d | undefined, transform?: Transform, testContainment?: boolean): number;
    setInvisible(invisible: boolean): void;
    addZClipPlanes(invisible: boolean, zLow?: number, zHigh?: number): void;
    appendPolygonClip(xyz: IndexedXYZCollection, insideFragments: GrowableXYZArray[], outsideFragments: GrowableXYZArray[], arrayCache: GrowableXYZArrayCache): void;
    static createConvexPolyface(convexMesh: Polyface | PolyfaceVisitor, result?: ConvexClipPlaneSet): {
        clipper: ConvexClipPlaneSet;
        volume: number;
    };
}